Rename EmailEvent.ts file to Email.event.ts
